requirements.txt
Django
razorpay
python-dotenv
dj-database-url
whitenoise
requests
Pillow
